CRUD separado por condomínio em Usuário

+++ src/main/java/br/com/onlares/dao/UsuarioDao.java
@@ -9,6 +9,7 @@ import javax.persistence.NoResultException;
 import javax.persistence.Query;
 
 import br.com.onlares.controller.UsuarioLogado;
+import br.com.onlares.model.Condominio;
 import br.com.onlares.model.Usuario;
 import br.com.onlares.util.MD5Hashing;
 
@@ -20,16 +21,11 @@ public class UsuarioDao {
 	@Inject
 	public UsuarioDao(EntityManager em, UsuarioLogado usuarioLogado) {
 		this.em = em;
-		this.condominioId = 1L;
-		//this.condominioId = (usuarioLogado != null ? usuarioLogado.getUsuario().getCondominio().getId() : null);
-		if (usuarioLogado != null) {
-			System.out.println("1="+usuarioLogado.getUsuario());	
-			if (usuarioLogado.getUsuario() != null) {
-				System.out.println("2="+usuarioLogado.getUsuario());	
-				if (usuarioLogado.getUsuario().getCondominio() != null) {
-					System.out.println("3="+usuarioLogado.getUsuario());	
-				}
-			}
+		if (usuarioLogado != null && usuarioLogado.getUsuario() != null
+				&& usuarioLogado.getUsuario().getCondominio() != null) {
+			this.condominioId = usuarioLogado.getUsuario().getCondominio().getId();
+		} else {
+			this.condominioId = null;
 		}
 	}
 	
@@ -38,9 +34,14 @@ public class UsuarioDao {
 		this(null, null); // para uso do CDI
 	}
 	
-	@SuppressWarnings("unchecked")
 	public List<Usuario> lista() {
-		return em.createQuery("select u from Usuario u").getResultList();
+		return em.createQuery("select u from Usuario u"
+				+ " where u.condominio.id = :condominioId", Usuario.class)
+				.setParameter("condominioId", condominioId).getResultList();
+	}
+
+	public List<Usuario> listaTodos() {
+		return em.createQuery("select u from Usuario u", Usuario.class).getResultList();
 	}
 	
 	public Usuario busca(Usuario usuario) {
@@ -61,36 +62,54 @@ public class UsuarioDao {
 	}
 	
 	public boolean loginValido(Usuario usuario) throws NoSuchAlgorithmException {
-		return !em.createQuery("select u from Usuario u where u.email = "
-			+ ":email and u.senha = :senha", Usuario.class)
+		return !em.createQuery("select u from Usuario u"
+			+ " where u.email = :email and u.senha = :senha", Usuario.class)
 			.setParameter("email", usuario.getEmail())
 			.setParameter("senha", MD5Hashing.convertStringToMd5(usuario.getSenha()))
 			.getResultList().isEmpty();
 	}
 	
 	public boolean existe(Usuario usuario) {
-		return !em.createQuery("select u from Usuario u where u.email = "
-			+ ":email", Usuario.class)
+		return !em.createQuery("select u from Usuario u"
+			+ " where u.email = :email"
+			+ " and u.condominio.id = :condominioId", Usuario.class)
 			.setParameter("email", usuario.getEmail())
+			.setParameter("condominioId", condominioId)
 			.getResultList().isEmpty();
 	}
 
 	public void adiciona(Usuario usuario) {
+		Condominio condominio = new Condominio();
+		condominio.setId(condominioId);
+		usuario.setCondominio(condominio);
 		em.getTransaction().begin();
 		em.persist(usuario);
 		em.getTransaction().commit();
 	}
 	
 	public void altera(Usuario usuario) {
-		em.getTransaction().begin();
-		em.merge(usuario);
-		em.getTransaction().commit();
+		if (mesmoCondominio(usuario)) {
+			em.getTransaction().begin();
+			em.merge(usuario);
+			em.getTransaction().commit();
+		}
 	}
 	
 	public void remove(Usuario usuario) {
-		em.getTransaction().begin();
-		em.remove(busca(usuario));
-		em.getTransaction().commit();
+		if (mesmoCondominio(usuario)) {
+			em.getTransaction().begin();
+			em.remove(busca(usuario));
+			em.getTransaction().commit();
+		}
+	}
+	
+	private boolean mesmoCondominio(Usuario usuario) {
+		if (usuario.getCondominio().getId().compareTo(condominioId) == 0) {
+			return true;
+		} else {
+			System.out.println("CONDOMÍNIOS DIFERENTES: " + usuario.getCondominio().getId() + " != " + condominioId);
+			return false;
+		}
 	}
 
 }
